+++ flow/__init__.py
@@ -1,2 +1,2 @@
 APP_NAME = "FlowMobile"
-APP_VERSION = "7.6.19"
+APP_VERSION = "8.0.0"